requirements.txt
pillow==10.1.0
pyinstaller==6.3.0 